+++ src/pyxel_mcp/screen_harness.py
@@ -55,6 +55,7 @@ def _flush_and_quit():
     print(json.dumps(_results))
     sys.stdout.flush()
     pyxel.quit()
+    os._exit(0)
 
 
 def _try_capture(fc, draw):